requirements.txt
pandas==3.0.2
python-dotenv==1.2.2
kaggle==2.1.0
